+++ src/nes/cpu/opcodes/cmp.rs
@@ -29,7 +29,7 @@ impl Immediate for Cmp {
 }
 
 impl FromAddress for Cmp {
-    fn from_address(cpu: &mut Cpu, mode: AddressMode) -> usize {
+    fn from_address(cpu: &mut Cpu, mode: AddressMode) -> u32 {
         let (src, extra_cycles) = cpu.translate_address(mode);
         let word = cpu.interconnect.read_word(src.to_u16());
 
@@ -38,9 +38,9 @@ impl FromAddress for Cmp {
         match mode {
             AddressMode::ZeroPage => 3,
             AddressMode::ZeroPageX => 4,
-            AddressMode::AbsoluteX => { 4 + (extra_cycles as usize) },
-            AddressMode::AbsoluteY => { 4 + (extra_cycles as usize) },
-            AddressMode::IndirectY => { 5 + (extra_cycles as usize) },
+            AddressMode::AbsoluteX => { 4 + (extra_cycles as u32) },
+            AddressMode::AbsoluteY => { 4 + (extra_cycles as u32) },
+            AddressMode::IndirectY => { 5 + (extra_cycles as u32) },
             _ => { panic!("unimplemented address mode {:?} for AND", mode); }
         }
     }
