+++ src/nes/cpu/opcodes/dec.rs
@@ -13,7 +13,7 @@ impl Dec {
 }
 
 impl FromAddress for Dec {
-    fn from_address(cpu: &mut Cpu, mode: AddressMode) -> usize {
+    fn from_address(cpu: &mut Cpu, mode: AddressMode) -> u32 {
         let (src, _) = cpu.translate_address(mode);
         let word = cpu.interconnect.read_word(src.to_u16());
         let result = Self::decrement(cpu, word);
