+++ src/nes/cpu/opcodes/sbc.rs
@@ -30,7 +30,7 @@ impl Sbc {
 }
 
 impl FromImmediate for Sbc {
-    fn from_immediate(cpu: &mut Cpu) -> usize {
+    fn from_immediate(cpu: &mut Cpu) -> u32 {
         let word = cpu.read_word_and_increment();
         Sbc::subtract_with_carry(cpu, word);
 
@@ -39,7 +39,7 @@ impl FromImmediate for Sbc {
 }
 
 impl FromAddress for Sbc {
-    fn from_address(cpu: &mut Cpu, mode: AddressMode) -> usize {
+    fn from_address(cpu: &mut Cpu, mode: AddressMode) -> u32 {
         let (src, extra_cycles) = cpu.translate_address(mode);
         let word = cpu.interconnect.read_word(src.to_u16());
 
@@ -48,9 +48,9 @@ impl FromAddress for Sbc {
         match mode {
             AddressMode::ZeroPage => 3,
             AddressMode::Absolute => 4,
-            AddressMode::AbsoluteX => { 4 + (extra_cycles as usize) },
-            AddressMode::AbsoluteY => { 4 + (extra_cycles as usize) },
-            AddressMode::IndirectY => { 5 + (extra_cycles as usize) },
+            AddressMode::AbsoluteX => { 4 + (extra_cycles as u32) },
+            AddressMode::AbsoluteY => { 4 + (extra_cycles as u32) },
+            AddressMode::IndirectY => { 5 + (extra_cycles as u32) },
             _ => { panic!("unimplemented address mode {:?} for SBC", mode); }
         }
     }
